Get rotations working for fixtures
