+++ Assets/Engine/Tile/Fixtures/WallFixtures/WallFixture.cs
@@ -34,17 +34,5 @@ namespace SS3D.Engine.Tiles
     abstract public class WallFixture : Fixture
     {
 
-        [SerializeField]
-        private Rotation rotation;
-
-        public void SetRotation(Rotation rotation)
-        {
-            this.rotation = rotation;
-        }
-
-        public Rotation GetRotation()
-        {
-            return rotation;
-        }
     }
 }
