+++ Assets/Engine/Tile/TileBase.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using SS3D.Engine.Tiles;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -11,5 +12,16 @@ public class TileBase : ScriptableObject
     public string genericType;
     public GameObject prefab;
 
-    
+    [SerializeField]
+    protected Rotation rotation = Rotation.North;
+
+    public void SetRotation(Rotation rotation)
+    {
+        this.rotation = rotation;
+    }
+
+    public Rotation GetRotation()
+    {
+        return rotation;
+    }
 }
